Extract update_tree_and_render_list to a simple mixin.
