+++ app/controllers/storage_controller.rb
@@ -4,6 +4,7 @@ class StorageController < ApplicationController
   include Mixins::GenericSessionMixin
   include Mixins::GenericShowMixin
   include Mixins::MoreShowActions
+  include Mixins::ExplorerPresenterMixin
 
   before_action :check_privileges
   before_action :get_session_data
@@ -518,19 +519,6 @@ class StorageController < ApplicationController
     presenter = rendering_objects
     update_tagging_partials(presenter)
     # update_title(presenter)
-    rebuild_toolbars(false, presenter)
-    handle_bottom_cell(presenter)
-
-    render :json => presenter.for_render
-  end
-
-  def update_tree_and_render_list(replace_trees)
-    @explorer = true
-    get_node_info(x_node)
-    presenter = rendering_objects
-    replace_explorer_trees(replace_trees, presenter)
-
-    presenter.update(:main_div, r[:partial => 'layouts/x_gtl'])
     rebuild_toolbars(false, presenter)
     handle_bottom_cell(presenter)
